+++ src/pages/tag/model/types.ts
@@ -1,0 +1,5 @@
+export interface Tag {
+  slug: string
+  name: string
+  url: string
+}
